fix: store available hashpack boolean in redux
by storing available hashpack boolean in useState, there
are some race condition when wallet modal might not mounted
and useEffect doesnt listen for already emitted event
so moving this logic to common place which is pangolin web3 provider
which will be available always
